refactor Menu component and make it dynamic
