+++ src/components/menu/Menu.css
@@ -1,0 +1,8 @@
+.menu {
+  font-size: 14px;
+  margin-top: 50px;
+}
+.menu-item {
+  list-style: none;
+  margin-top: 10px;
+}
